styling and re-organize AvailablePets component
